Move NavBar to common
