+++ frontend/src/TopBar.js
@@ -11,7 +11,8 @@ import {routeTo} from './helpers/routes'
 import Search from './screens/Blockchain/BlockchainHeader/Search'
 import seizaLogoDesktop from './assets/icons/logo-seiza.svg'
 import seizaLogoMobile from './assets/icons/seiza-symbol.svg'
-import {NavLinks, MobileNavLinks, Link} from './components/visual'
+import {NavLinks, MobileNavLinks} from '@/components/common/Navbar'
+import {Link} from '@/components/visual'
 
 import {useIsMobile} from '@/components/hooks/useBreakpoints'
